requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
sqlalchemy>=2.0.28
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-multipart>=0.0.9
requests>=2.31.0
psutil>=5.9.8
packaging>=24.0
